fix: displaying the error in a different way
